+++ capytaine/__init__.py
@@ -11,6 +11,9 @@ from capytaine.meshes.geometry import Axis, Plane, xOz_Plane, yOz_Plane, xOy_Pla
 from capytaine.meshes.meshes import Mesh
 from capytaine.meshes.collections import CollectionOfMeshes
 from capytaine.meshes.symmetric import ReflectionSymmetricMesh, TranslationalSymmetricMesh, AxialSymmetricMesh
+from capytaine.meshes.predefined.cylinders import mesh_disk, mesh_horizontal_cylinder, mesh_vertical_cylinder
+from capytaine.meshes.predefined.spheres import mesh_sphere
+from capytaine.meshes.predefined.rectangles import mesh_rectangle, mesh_parallelepiped
 
 from capytaine.bodies.bodies import FloatingBody
 from capytaine.bodies.dofs import rigid_body_dofs