+++ docs/user_manual/examples/multibody.py
@@ -4,24 +4,28 @@
 import capytaine as cpt
 
 # Define the first body
-sphere = cpt.Sphere(radius=1.0, center=(0, 0, -2.0),
-                ntheta=20, nphi=20,
-                name="sphere_1")
+sphere = cpt.FloatingBody(
+        cpt.mesh_sphere(radius=1.0, center=(0, 0, -2.0),
+                        resolution=(20, 20)),
+        name="sphere_1")
 sphere.add_translation_dof(name="Surge")
 sphere.add_translation_dof(name="Heave")
 
 # Define the second body
-other_sphere = cpt.Sphere(radius=0.5, center=(-2, -3, -1),
-                      ntheta=20, nphi=20,
-                      name="sphere_2")
+other_sphere = cpt.FloatingBody(
+        cpt.mesh_sphere(radius=0.5, center=(-2, -3, -1),
+                        resolution=(20, 20)),
+        name="sphere_2")
 other_sphere.add_translation_dof(name="Surge")
 other_sphere.add_translation_dof(name="Heave")
 
 # Define the third body
-cylinder = cpt.HorizontalCylinder(length=5.0, radius=1.0,
-                              center=(1.5, 3.0, -3.0),
-                              nx=20, nr=3, ntheta=20,
-                              name="cylinder")
+cylinder = cpt.FloatingBody(
+        cpt.mesh_horizontal_cylinder(
+            length=5.0, radius=1.0,
+            center=(1.5, 3.0, -3.0),
+            resolution=(20, 20, 3)),
+        name="cylinder")
 cylinder.add_translation_dof(name="Surge")
 cylinder.add_translation_dof(name="Heave")